can view applicant infoa and added validation in delete user

+++ src/app/admin/admin-users/admin-applicant-list/admin-applicant-list.component.ts
@@ -3,6 +3,8 @@ import { AdminUserService } from '../../../core';
 import { MatTableDataSource, MatTable } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog, MatDialogConfig, MatDialogRef } from '@angular/material/dialog';
+import { AdminEmployeeViewComponent } from '../admin-employee-view/admin-employee-view.component'
 
 @Component({
   selector: 'app-admin-applicant-list',
@@ -30,7 +32,8 @@ export class AdminApplicantListComponent implements OnInit {
 
   constructor(
     private adminUserservice: AdminUserService,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private matDialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -42,15 +45,36 @@ export class AdminApplicantListComponent implements OnInit {
   }
 
   deleteUser(id){
-    this.adminUserservice.deleteUser(id).subscribe( result =>{
-      const index = this.dataSource.data.indexOf(id);
+    const dialogRef = this.matDialog.open(
+      ApplicantDeleteDialog,{
+        width: '300px',
+      }
+    );
 
-      this.snackBar.open('User Deleted', 'close');
-      this.dataSource.data.splice(index, 1);
-      this.dataSource._updateChangeSubscription(); 
-      this.dataSource.paginator = this.paginator; 
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.adminUserservice.deleteUser(id).subscribe( result =>{
+        const index = this.dataSource.data.indexOf(id);
+        this.snackBar.open('User Deleted', 'close');
+        this.dataSource.data.splice(index, 1);
+        this.dataSource._updateChangeSubscription(); 
+        this.dataSource.paginator = this.paginator; 
+       })
+      }
+    });
+  }
 
-    })
+  viewUserInfo(data){
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.id = 'create-user';
+    dialogConfig.height = '90%';
+    dialogConfig.width = '1000px';
+    dialogConfig.data = { data:data };
+    const modalDialog = this.matDialog.open(
+      AdminEmployeeViewComponent,
+      dialogConfig
+    );
 
   }
 
@@ -60,4 +84,21 @@ export class AdminApplicantListComponent implements OnInit {
 
 
 
+}
+
+@Component({
+  selector: 'dialog-applicant-delete',
+  templateUrl: './admin-applicant-delete.component.html',
+  styleUrls: ['./admin-applicant-list.component.scss'],
+})
+export class ApplicantDeleteDialog {
+
+  constructor(
+    public dialogRef: MatDialogRef<ApplicantDeleteDialog>
+    ) {}
+
+  onNoClick(): void {
+    this.dialogRef.close();
+  }
+ 
 }
